+++ frontend/src/classes/Character.js
@@ -1,5 +1,5 @@
 
-import {CharacterApi, PlayerApi, ServiceApi} from "neucore-js-client";
+import {CharacterApi, PlayerApi} from "neucore-js-client";
 
 export default class Character {
 
@@ -10,8 +10,9 @@ export default class Character {
     /**
      * @param {int} characterId
      * @param {function} [callback]
+     * @param {string} [successMessage]
      */
-    updateCharacter(characterId, callback) {
+    updateCharacter(characterId, callback, successMessage) {
         const self = this;
         new CharacterApi().update(characterId, function(error, data, response) {
             if (error) { // usually 403 (from Core) or 503 (ESI down)
@@ -27,51 +28,12 @@ export default class Character {
                     'info'
                 );
             } else {
-                self.vm.message('Updated character.', 'success');
+                self.vm.message(successMessage || 'Character updated.', 'success');
             }
             if (typeof callback === typeof Function) {
                 callback();
             }
         });
-    }
-
-    /**
-     * Updates all characters from ESI, groups and service accounts of the player.
-     *
-     * @param {object} player
-     * @param {function} [callback]
-     */
-    updatePlayer(player, callback) {
-        const self = this;
-        const characters = [...player.characters];
-
-        function updateCharacters() {
-            if (characters.length > 0) {
-                const id = characters[0].id;
-                characters.splice(0, 1);
-                self.updateCharacter(id, function() {
-                    updateCharacters();
-                });
-            } else {
-                charactersUpdateComplete()
-            }
-        }
-
-        function charactersUpdateComplete() {
-            new ServiceApi().serviceUpdateAllAccounts(player.id, (error, data) => {
-                if (!error) {
-                    self.vm.message(`Updated ${data.length} service accounts.`, 'success', 2500);
-                }
-                if (typeof callback === typeof Function) {
-                    callback();
-                }
-                if (player.id === self.vm.$root.player.id) {
-                    self.vm.emitter.emit('playerChange');
-                }
-            })
-        }
-
-        updateCharacters();
     }
 
     /**